--- examples/website/tagmap/tagmap-layer/index.js
@@ -1,1 +1,0 @@
-export {default} from './tagmap-layer';
