+++ examples/website/text/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "deckgl-examples-tagmap",
+  "version": "0.0.0",
+  "private": true,
+  "license": "MIT",
+  "scripts": {
+    "start": "vite --open",
+    "start-local": "vite --config ../../vite.config.local.mjs",
+    "build": "vite build"
+  },
+  "dependencies": {
+    "d3-scale": "^2.0.0",
+    "d3-request": "^1.0.6",
+    "deck.gl": "^8.8.0",
+    "maplibre-gl": "^2.4.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-map-gl": "^7.0.0",
+    "tagmap.js": "^1.1.1"
+  },
+  "devDependencies": {
+    "typescript": "^4.6.0",
+    "vite": "^4.0.0"
+  }
+}
