+++ examples/website/text/tagmap-layer/index.js
@@ -1,0 +1,1 @@
+export {default} from './tagmap-layer';
